Extract BackendAdmissionConfig into admission/config.rs
