+++ crates/defra-agent/src/admission/mod.rs
@@ -6,6 +6,8 @@ mod permit;
 mod registry;
 mod persistence;
 
+pub(crate) use config::{backend_admission_configs_from_backends, BackendAdmissionConfig};
+
 use std::collections::{HashMap, HashSet};
 use std::future::Future;
 use std::sync::atomic::{AtomicBool, AtomicU64, AtomicUsize, Ordering};
@@ -21,64 +23,8 @@ use rig::streaming::StreamingCompletionResponse;
 use tokio::sync::{OwnedSemaphorePermit, Semaphore};
 
 use self::stream_guard::{hold_stream_guard, StreamGuardLifecycle};
-use crate::backend_registry::InferenceBackend;
 use crate::graphql::escape_graphql_string;
 use crate::watcher::AgentRequest;
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub(crate) struct BackendAdmissionConfig {
-    pub(crate) backend_id: String,
-    pub(crate) max_concurrent: usize,
-    pub(crate) max_queue_depth: usize,
-    pub(crate) enabled: bool,
-    pub(crate) probe_status: String,
-    pub(crate) config_fingerprint: String,
-}
-
-impl BackendAdmissionConfig {
-    pub(crate) fn from_backend(backend: &InferenceBackend) -> Result<Self> {
-        if backend.max_concurrent < 1 {
-            anyhow::bail!(
-                "backend {} has invalid max_concurrent {}; expected >= 1",
-                backend.backend_id,
-                backend.max_concurrent
-            );
-        }
-        if backend.max_queue_depth < 0 {
-            anyhow::bail!(
-                "backend {} has invalid max_queue_depth {}; expected >= 0",
-                backend.backend_id,
-                backend.max_queue_depth
-            );
-        }
-
-        Ok(Self {
-            backend_id: backend.backend_id.clone(),
-            max_concurrent: backend.max_concurrent as usize,
-            max_queue_depth: backend.max_queue_depth as usize,
-            enabled: backend.enabled,
-            probe_status: backend.probe_status.clone(),
-            config_fingerprint: format!("{backend:?}"),
-        })
-    }
-
-    pub(crate) fn is_available(&self) -> bool {
-        self.enabled && self.probe_status == "healthy"
-    }
-}
-
-pub(crate) fn backend_admission_configs_from_backends<'a>(
-    backends: impl IntoIterator<Item = &'a InferenceBackend>,
-) -> Result<HashMap<String, BackendAdmissionConfig>> {
-    let mut configs = HashMap::new();
-    for backend in backends {
-        configs.insert(
-            backend.backend_id.clone(),
-            BackendAdmissionConfig::from_backend(backend)?,
-        );
-    }
-    Ok(configs)
-}
 
 #[derive(Clone)]
 pub(crate) struct AdmittedCompletionClient<C> {
